Lots of tricky things here
Made sure all the migration to slugs within the settings is complete. I
think I still need to remove it from the database as its own field.

Fixed an events recovery bug, where the session needs to be captured and
"recreated"... aka restore the session token to the context.

Along the context lines, make use of context.WithoutCancel() to continue
passing the context around with all its values but in a way that can
persist across the http request ending.

Also made a few more things dynamic in the profile and migrated the
callsign finder to the finder interface. This way it can get cached.